Remove outdated Bluetooth and RF24L01 example scripts; add UDP communication examples for ESP32. python and c++ with the correct updated communication method.
